backend/requirements.txt
flask
flask-cors
flask-babel
requests
pandas
Pillow
cloudinary
pymongo
python-dotenv
dnspython
xgboost
scikit-learn
lime
langchain
langchain-huggingface
langchain-community
langchain-groq
tensorflow
keras
